refactor: implement useExtensionToastNotification hook for enhanced toast notifications and update SocketGraphComponent to utilize the new hook, improving user engagement
